+++ Changelog.md
@@ -4,6 +4,7 @@
 - Fix: Nether portals can replace reinforced blocks when generating
 - Fix: Can't place blocks on blocks locked by a Sonic Security System
 - Fix: Interacting with blocks while holding a Sonic Security System may sometimes place it for a short while
+- Fix: A Sentry's name does not show in the Sentry Remote Access Tool
 
 --------------------------Changelog for v1.9.3.1 of SecurityCraft--------------------------